+++ cmd/validator/depositdata/run.go
@@ -1,4 +1,4 @@
-// Copyright © 2019, 2020 Weald Technology Trading
+// Copyright © 2019 - 2024 Weald Technology Trading.
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
@@ -14,7 +14,9 @@
 package depositdata
 
 import (
-	"github.com/pkg/errors"
+	"context"
+	"errors"
+
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 )
@@ -23,7 +25,7 @@ import (
 func Run(cmd *cobra.Command) (string, error) {
 	dataIn, err := input()
 	if err != nil {
-		return "", errors.Wrap(err, "failed to obtain input")
+		return "", errors.Join(errors.New("failed to set up command"), err)
 	}
 
 	// Further errors do not need a usage report.
@@ -31,7 +33,12 @@ func Run(cmd *cobra.Command) (string, error) {
 
 	dataOut, err := process(dataIn)
 	if err != nil {
-		return "", errors.Wrap(err, "failed to process")
+		switch {
+		case errors.Is(err, context.DeadlineExceeded):
+			return "", errors.New("operation timed out; try increasing with --timeout option")
+		default:
+			return "", errors.Join(errors.New("failed to process"), err)
+		}
 	}
 
 	if viper.GetBool("quiet") {
@@ -40,7 +47,7 @@ func Run(cmd *cobra.Command) (string, error) {
 
 	results, err := output(dataOut)
 	if err != nil {
-		return "", errors.Wrap(err, "failed to obtain output")
+		return "", errors.Join(errors.New("failed to obtain output"), err)
 	}
 
 	return results, nil