+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+dev:
+  - provide better error message on context deadlline exceeded
+
 1.35.2:
   - update dependencies